+++ packages/web-pkg/src/types.ts
@@ -7,3 +7,14 @@ export interface AncestorMetaDataValue {
 }
 
 export type AncestorMetaData = Record<string, AncestorMetaDataValue>
+
+export type FederatedUser = {
+  display_name: string
+  idp: string
+  mail: string
+  user_id: string
+}
+
+export type FederatedConnection = FederatedUser & {
+  id: string
+}
